+++ BovineLabs.Core/Extensions/WorldExtensions.cs
@@ -67,5 +67,21 @@ namespace BovineLabs.Core.Extensions
         {
             return (world.Flags & WorldFlags.GameServer) == WorldFlags.GameServer;
         }
+
+        /// <summary> Check if a world is the editor world. </summary>
+        /// <param name="world"> A <see cref="WorldUnmanaged" /> instance </param>
+        /// <returns> </returns>
+        public static bool IsEditorWorld(this World world)
+        {
+            return (world.Flags & WorldFlags.Editor) == WorldFlags.Editor;
+        }
+
+        /// <summary> Check if an unmanaged world is the editor world. </summary>
+        /// <param name="world"> A <see cref="WorldUnmanaged" /> instance </param>
+        /// <returns> If it's an editor world. </returns>
+        public static bool IsEditorWorld(this WorldUnmanaged world)
+        {
+            return (world.Flags & WorldFlags.Editor) == WorldFlags.Editor;
+        }
     }
 }