+++ BovineLabs.Core/Extensions/WorldUnmanagedExtensions.cs
@@ -4,7 +4,9 @@
 
 namespace BovineLabs.Core.Extensions
 {
+    using Unity.Collections;
     using Unity.Entities;
+    using Unity.Jobs;
 
     public static class WorldUnmanagedExtensions
     {
@@ -12,6 +14,16 @@ namespace BovineLabs.Core.Extensions
         {
             var typeIndex = TypeManager.GetSystemTypeIndex<T>();
             return world.GetExistingUnmanagedSystem(typeIndex) != SystemHandle.Null;
+        }
+
+        public static unsafe void GetAllSystemDependencies(this WorldUnmanaged world, NativeList<JobHandle> dependencies)
+        {
+            using var e = world.GetImpl().m_SystemStatePtrMap.GetEnumerator();
+            while (e.MoveNext())
+            {
+                var systemState = (SystemState*)e.Current.Value;
+                dependencies.Add(systemState->m_JobHandle);
+            }
         }
     }
 }